feat: Revamped action system as a new addon
