+++ addons/geary-actions/src/commonMain/kotlin/com/mineinabyss/geary/actions/Action.kt
@@ -1,0 +1,5 @@
+package com.mineinabyss.geary.actions
+
+interface Action {
+    fun ActionGroupContext.execute()
+}
