+++ addons/geary-actions/src/commonMain/kotlin/com/mineinabyss/geary/actions/ActionsCancelledException.kt
@@ -1,0 +1,3 @@
+package com.mineinabyss.geary.actions
+
+class ActionsCancelledException : Exception()
