+++ addons/geary-actions/src/commonMain/kotlin/com/mineinabyss/geary/actions/Condition.kt
@@ -1,0 +1,5 @@
+package com.mineinabyss.geary.actions
+
+interface Condition {
+    fun ActionGroupContext.execute(): Boolean
+}
