+++ addons/geary-actions/src/commonMain/kotlin/com/mineinabyss/geary/actions/GearyActions.kt
@@ -1,0 +1,18 @@
+package com.mineinabyss.geary.actions
+
+import com.mineinabyss.geary.actions.event_binds.bindEntityObservers
+import com.mineinabyss.geary.addons.GearyPhase
+import com.mineinabyss.geary.addons.dsl.GearyAddonWithDefault
+import com.mineinabyss.geary.modules.geary
+
+class GearyActions {
+    companion object : GearyAddonWithDefault<GearyActions> {
+        override fun default() = GearyActions()
+
+        override fun GearyActions.install() {
+            geary.run {
+                bindEntityObservers()
+            }
+        }
+    }
+}
