+++ addons/geary-actions/src/commonMain/kotlin/com/mineinabyss/geary/actions/event_binds/EventBind.kt
@@ -1,0 +1,10 @@
+package com.mineinabyss.geary.actions.event_binds
+
+import com.mineinabyss.geary.serialization.serializers.SerializableComponentId
+import com.mineinabyss.geary.serialization.serializers.SerializedComponents
+
+class EventBind(
+    val event: SerializableComponentId,
+    val involving: List<SerializableComponentId> = listOf(),
+    val emit: List<SerializedComponents>,
+)
