+++ addons/geary-prefabs/src/commonMain/kotlin/com/mineinabyss/geary/prefabs/configuration/components/ReEmitEvent.kt
@@ -1,6 +1,7 @@
 package com.mineinabyss.geary.prefabs.configuration.components
 
 import com.mineinabyss.geary.datatypes.ComponentId
+import com.mineinabyss.geary.serialization.serializers.SerializableComponentId
 
 data class ReEmitEvent(
     val findByRelationKind: SerializableComponentId,
