+++ addons/geary-serialization/src/commonMain/kotlin/com/mineinabyss/geary/serialization/serializers/SerializableComponentId.kt
@@ -1,9 +1,8 @@
-package com.mineinabyss.geary.prefabs.configuration.components
+package com.mineinabyss.geary.serialization.serializers
 
 import com.mineinabyss.geary.datatypes.ComponentId
 import com.mineinabyss.geary.helpers.componentId
 import com.mineinabyss.geary.serialization.serializableComponents
-import com.mineinabyss.geary.serialization.serializers.PolymorphicListAsMapSerializer
 import kotlinx.serialization.KSerializer
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.descriptors.PrimitiveKind
